Change Plotly Vis to be Main Visualizing Class. Remove Outdated Vis Code.
Previous visualizer.py is now part of plotly visualizer.
